+++ tests/Suru.Tests/ArrayTests.cs
@@ -7,6 +7,5 @@ public class ArrayTests(CompiledFixtures fixtures) : IntegrationTestBase
 
     [Fact]
     public void Array_LenAtSetAddSlice()
-        // => Assert.Equal("3\n10\n30\n99\n4\n40\n2\n99\n4\n3\n4\n3 three\n", Run(_exe));
-        => Assert.Equal("3\n10\n30\n99\n4\n40\n2\n99\n", Run(_exe));
+        => Assert.Equal("3\n10\n30\n99\n4\n40\n2\n99\n4\n10\n777\n10\n2\nhello\n", Run(_exe));
 }